+++ doc/source/doxygen-docs/changelog.md
@@ -18,6 +18,7 @@
       - Texture handling refactored into a new class mrpt::opengl::Texture
       - Enabled GL MipMap texture generation.
       - Triangles shaders now also account for shininess of material to generate specular reflections. See new method mrpt::opengl::CRenderizable::materialShininess()
+      - mrpt::opengl::TLightParameters has been modified to make it more consistent: just one light color and three floats (diffuse, ambient, specular).
 - BUG FIXES:
     - Fix unrealistic odometry error simulation in mrpt::kinematics::CVehicleSimulVirtualBase (it affected the app GridmapNavSimul).